+++ packages/Filter/stories/Showcase/Showcase.js
@@ -72,7 +72,7 @@ export default function App() {
 
   return (
     <React.Fragment>
-      <Filter {...getFilterProps()} columns={columnsSettings} data={data}>
+      <Filter {...getFilterProps()} columns={columnsSettings} data={data} maxFiltersAllowed={5}>
         {filters.map((filter, index) => (
           <Filter.Item
             {...getFilterItemProps()}