+++ packages/L10n/stories/L10n.examples.stories.js
@@ -3,7 +3,8 @@ import { getStoryName } from "storybook/storyTree";
 import ExampleStory from "storybook/components/ExampleStory";
 import { exampleStoryParameters } from "storybook/assets/storyParameters";
 import FakeAppWithoutContext from "./examples/FakeAppWithoutContext";
-import FakeAppWithLocales from "./examples/withLocales/FakeAppWithLocales";
+import FakeAppWithLocales from "./examples/FakeAppWithLocales";
+import FakeAppWithInterpolation from "./examples/FakeAppWithInterpolation";
 
 import L10n from "../src";
 
@@ -13,6 +14,8 @@ export default {
   title: `${storyName}/Examples`,
   component: L10n,
 };
+
+//
 
 export const FakeAppWithLocalesStory = () => (
   <ExampleStory component="L10n" storyName="App with Locales" fileName="examples/withLocales/FakeAppWithLocales.js">
@@ -25,6 +28,8 @@ FakeAppWithLocalesStory.story = {
   parameters: exampleStoryParameters,
 };
 
+//
+
 export const FakeAppWithoutContextStory = () => (
   <ExampleStory component="L10n" storyName="App without Context" fileName="examples/FakeAppWithoutContext.js">
     <FakeAppWithoutContext />
@@ -33,5 +38,18 @@ export const FakeAppWithoutContextStory = () => (
 
 FakeAppWithoutContextStory.story = {
   name: "App without Context",
+  parameters: exampleStoryParameters,
+};
+
+//
+
+export const FakeAppWithInterpolationStory = () => (
+  <ExampleStory component="L10n" storyName="App with Interpolation" fileName="examples/FakeAppWithInterpolation.js">
+    <FakeAppWithInterpolation />
+  </ExampleStory>
+);
+
+FakeAppWithInterpolationStory.story = {
+  name: "App with Interpolation",
   parameters: exampleStoryParameters,
 };