+++ packages/Toast/stories/examples/ControlledToastExample.js
@@ -5,6 +5,12 @@ import Toast from "../../src";
 
 export default function ControlledToast() {
   const [isOpen, setIsOpen] = React.useState(true);
+  const [isAutoCloseToastOpen, setIsAutoCloseToastOpen] = React.useState(true);
+
+  function handleClick() {
+    setIsAutoCloseToastOpen(prev => !prev);
+  }
+
   return (
     <L10n locale="en">
       <Toast
@@ -16,6 +22,19 @@ export default function ControlledToast() {
       >
         Controlled toast component
       </Toast>
+      <Toast
+        hasCloseButton
+        isOpen={isAutoCloseToastOpen}
+        canAutoClose
+        onClose={() => {
+          setIsAutoCloseToastOpen(false);
+        }}
+      >
+        Controlled toast component with auto close
+      </Toast>
+      <button type="button" onClick={handleClick}>
+        Toggle Autoclose toast
+      </button>
     </L10n>
   );
 }