+++ packages/DialogActions/CHANGELOG.md
@@ -1,5 +1,54 @@
 # Changelog
 
+## 1.0.15
+
+### Patch Changes
+
+- Updated dependencies [ff8eae9]
+  - @paprika/l10n@1.1.13
+  - @paprika/button@1.1.2
+
+## 1.0.15-next.0
+
+### Patch Changes
+
+- Updated dependencies [ff8eae9]
+  - @paprika/l10n@1.1.13-next.0
+  - @paprika/button@1.1.2-next.0
+
+## 1.0.14
+
+### Patch Changes
+
+- Updated dependencies [b20c1d4]
+  - @paprika/l10n@1.1.12
+
+## 1.0.14-next.0
+
+### Patch Changes
+
+- Updated dependencies [b20c1d4]
+  - @paprika/l10n@1.1.12-next.0
+
+## 1.0.13
+
+### Patch Changes
+
+- @paprika/button@1.1.1
+
+## 1.0.13-next.0
+
+### Patch Changes
+
+- @paprika/button@1.1.1-next.0
+
+## 1.0.12
+
+### Patch Changes
+
+- Updated dependencies [ae0d6a0]
+  - @paprika/l10n@1.1.11
+
 ## 1.0.12-next.0
 
 ### Patch Changes