+++ packages/DynamicHyperlinkTransformer/CHANGELOG.md
@@ -1,5 +1,48 @@
 # @paprika/dynamic-hyperlink-transformer
 
+## 1.1.17
+
+### Patch Changes
+
+- Updated dependencies [ff8eae9]
+  - @paprika/l10n@1.1.13
+  - @paprika/stylers@1.0.1
+  - @paprika/tabs@2.2.6
+
+## 1.1.17-next.0
+
+### Patch Changes
+
+- Updated dependencies [ff8eae9]
+  - @paprika/l10n@1.1.13-next.0
+  - @paprika/stylers@1.0.1-next.0
+  - @paprika/tabs@2.2.6-next.0
+
+## 1.1.16
+
+### Patch Changes
+
+- Updated dependencies [b20c1d4]
+  - @paprika/l10n@1.1.12
+
+## 1.1.16-next.0
+
+### Patch Changes
+
+- Updated dependencies [b20c1d4]
+  - @paprika/l10n@1.1.12-next.0
+
+## 1.1.15
+
+### Patch Changes
+
+- Updated dependencies [50fc8d6]
+- Updated dependencies [f2de3be]
+- Updated dependencies [ae0d6a0]
+- Updated dependencies [17bb9de]
+  - @paprika/tabs@2.2.5
+  - @paprika/l10n@1.1.11
+
 ## 1.1.15-next.1
 
 ### Patch Changes