+++ packages/Popover/CHANGELOG.md
@@ -1,5 +1,23 @@
 # Changelog
 
+## 1.2.4
+
+### Patch Changes
+
+- Updated dependencies [ff8eae9]
+  - @paprika/stylers@1.0.1
+  - @paprika/helpers@2.1.3
+  - @paprika/raw-button@1.0.5
+
+## 1.2.4-next.0
+
+### Patch Changes
+
+- Updated dependencies [ff8eae9]
+  - @paprika/stylers@1.0.1-next.0
+  - @paprika/helpers@2.1.3-next.0
+  - @paprika/raw-button@1.0.5-next.0
+
 ## 1.2.3
 
 ### Patch Changes