+++ packages/StatusTracker/CHANGELOG.md
@@ -1,5 +1,83 @@
 # Changelog
 
+## 1.1.10
+
+### Patch Changes
+
+- Updated dependencies [ff8eae9]
+  - @paprika/l10n@1.1.13
+  - @paprika/stylers@1.0.1
+  - @paprika/helpers@2.1.3
+  - @paprika/button@1.1.2
+  - @paprika/overflow-menu@1.1.3
+  - @paprika/popover@1.2.4
+  - @paprika/raw-button@1.0.5
+
+## 1.1.10-next.0
+
+### Patch Changes
+
+- Updated dependencies [ff8eae9]
+  - @paprika/l10n@1.1.13-next.0
+  - @paprika/stylers@1.0.1-next.0
+  - @paprika/helpers@2.1.3-next.0
+  - @paprika/overflow-menu@1.1.3-next.0
+  - @paprika/button@1.1.2-next.0
+  - @paprika/popover@1.2.4-next.0
+  - @paprika/raw-button@1.0.5-next.0
+
+## 1.1.9
+
+### Patch Changes
+
+- Updated dependencies [b20c1d4]
+  - @paprika/l10n@1.1.12
+  - @paprika/overflow-menu@1.1.2
+
+## 1.1.9-next.0
+
+### Patch Changes
+
+- Updated dependencies [b20c1d4]
+  - @paprika/l10n@1.1.12-next.0
+  - @paprika/overflow-menu@1.1.2-next.0
+
+## 1.1.8
+
+### Patch Changes
+
+- @paprika/button@1.1.1
+- @paprika/overflow-menu@1.1.1
+
+## 1.1.8-next.0
+
+### Patch Changes
+
+- @paprika/button@1.1.1-next.0
+- @paprika/overflow-menu@1.1.1-next.0
+
+## 1.1.7
+
+### Patch Changes
+
+- Updated dependencies [11ca2a7]
+  - @paprika/overflow-menu@1.1.0
+
+## 1.1.7-next.0
+
+### Patch Changes
+
+- Updated dependencies [11ca2a7]
+  - @paprika/overflow-menu@1.1.0-next.0
+
+## 1.1.6
+
+### Patch Changes
+
+- Updated dependencies [ae0d6a0]
+  - @paprika/l10n@1.1.11
+  - @paprika/overflow-menu@1.0.20
+
 ## 1.1.6-next.0
 
 ### Patch Changes